requirements.txt
cbor2==5.4.6
certifi==2022.12.7
charset-normalizer==3.1.0
crytic-compile==0.3.0
idna==3.4
packaging==23.0
prettytable==3.6.0
py-solc-x==1.1.1
pycryptodome==3.17
requests==2.28.2
semantic-version==2.10.0
slither-analyzer==0.9.2
urllib3==1.26.15
wcwidth==0.2.6
